Move some rendering code for traffic sim things from map_gui to game, the only place it's needed
