+++ apps/game/src/info/building.rs
@@ -1,7 +1,6 @@
 use std::collections::BTreeMap;
 
 use geom::{Angle, Circle, Distance, Speed, Time};
-use map_gui::render::DrawPedestrian;
 use map_model::{BuildingID, LaneID, OffstreetParking, Traversable, SIDEWALK_THICKNESS};
 use sim::{DrawPedestrianInput, PedestrianID, PersonID, TripResult, VehicleType};
 use synthpop::TripMode;
@@ -9,6 +8,7 @@ use widgetry::{Color, EventCtx, Line, Text, TextExt, Widget};
 
 use crate::app::App;
 use crate::info::{header_btns, make_table, make_tabs, Details, Tab};
+use crate::render::DrawPedestrian;
 
 pub fn info(ctx: &mut EventCtx, app: &App, details: &mut Details, id: BuildingID) -> Widget {
     Widget::custom_col(vec![
