+++ apps/game/src/render/car.rs
@@ -1,11 +1,12 @@
 use geom::{Angle, ArrowCap, Distance, PolyLine, Polygon, Pt2D, Ring, Tessellation};
+use map_gui::colors::ColorScheme;
+use map_gui::render::{DrawOptions, Renderable, OUTLINE_THICKNESS};
+use map_gui::{AppLike, ID};
 use map_model::{Map, TurnType};
 use sim::{CarID, CarStatus, DrawCarInput, Intent, Sim, VehicleType};
 use widgetry::{Color, Drawable, GeomBatch, GfxCtx, Line, Prerender, Text};
 
-use crate::colors::ColorScheme;
-use crate::render::{grey_out_unhighlighted_people, DrawOptions, Renderable, OUTLINE_THICKNESS};
-use crate::{AppLike, ID};
+use crate::render::grey_out_unhighlighted_people;
 
 const CAR_WIDTH: Distance = Distance::const_meters(1.75);
