+++ map_gui/src/render/mod.rs
@@ -1,32 +1,22 @@
-//! Render static and dynamic map elements.
+//! Render static map elements.
 
 use geom::{Distance, Pt2D, Tessellation};
-use map_model::{IntersectionID, Map, NORMAL_LANE_THICKNESS, SIDEWALK_THICKNESS};
-use sim::{DrawCarInput, PersonID, Sim, VehicleType};
-use widgetry::{Color, GfxCtx, Prerender};
+use map_model::{IntersectionID, Map};
+use widgetry::GfxCtx;
 
-use crate::colors::ColorScheme;
-pub use crate::render::agents::{AgentCache, UnzoomedAgents};
 pub use crate::render::area::DrawArea;
-use crate::render::bike::DrawBike;
 pub use crate::render::building::DrawBuilding;
-use crate::render::car::DrawCar;
 pub use crate::render::intersection::{calculate_corners, DrawIntersection};
 pub use crate::render::map::DrawMap;
-pub use crate::render::pedestrian::{DrawPedCrowd, DrawPedestrian};
 pub use crate::render::turn::DrawMovement;
 use crate::{AppLike, ID};
 
-mod agents;
 mod area;
-mod bike;
 mod building;
-mod car;
 mod intersection;
 mod lane;
 mod map;
 mod parking_lot;
-mod pedestrian;
 mod road;
 pub mod traffic_signal;
 mod transit_stop;
@@ -54,20 +44,6 @@ pub trait Renderable {
     fn contains_pt(&self, pt: Pt2D, map: &Map) -> bool;
 }
 
-fn draw_vehicle(
-    input: DrawCarInput,
-    map: &Map,
-    sim: &Sim,
-    prerender: &Prerender,
-    cs: &ColorScheme,
-) -> Box<dyn Renderable> {
-    if input.id.vehicle_type == VehicleType::Bike {
-        Box::new(DrawBike::new(input, map, sim, prerender, cs))
-    } else {
-        Box::new(DrawCar::new(input, map, sim, prerender, cs))
-    }
-}
-
 /// Control how the map is drawn.
 pub struct DrawOptions {
     /// Don't draw the current traffic signal state.
@@ -84,28 +60,4 @@ impl DrawOptions {
             label_buildings: false,
         }
     }
-}
-
-pub fn unzoomed_agent_radius(vt: Option<VehicleType>) -> Distance {
-    // Lane thickness is a little hard to see, so double it. Most of the time, the circles don't
-    // leak out of the road too much.
-    if vt.is_some() {
-        4.0 * NORMAL_LANE_THICKNESS
-    } else {
-        4.0 * SIDEWALK_THICKNESS
-    }
-}
-
-/// If the sim has highlighted people, then fade all others out.
-fn grey_out_unhighlighted_people(color: Color, person: &Option<PersonID>, sim: &Sim) -> Color {
-    if let Some(ref highlighted) = sim.get_highlighted_people() {
-        if person
-            .as_ref()
-            .map(|p| !highlighted.contains(p))
-            .unwrap_or(false)
-        {
-            return color.tint(0.5);
-        }
-    }
-    color
 }
